+++ src/main/java/me/confuser/banmanager/commands/WarnCommand.java
@@ -163,7 +163,7 @@ public class WarnCommand extends AutoCompleteNameTabCommand<BanManager> {
 
         try {
           actionCommands = plugin.getConfiguration().getWarningActions()
-                                 .getCommand((int) plugin.getPlayerWarnStorage().getPointsCount(player));
+                                 .getCommand(plugin.getPlayerWarnStorage().getPointsCount(player));
         } catch (SQLException e) {
           e.printStackTrace();
           return;
@@ -184,7 +184,7 @@ public class WarnCommand extends AutoCompleteNameTabCommand<BanManager> {
                                            .replace("[playerId]", player.getUUID().toString())
                                            .replace("[actor]", actor.getName())
                                            .replace("[reason]", warning.getReason())
-                                           .replace("[points]", Integer.toString(parser.getPoints()));
+                                           .replace("[points]", Double.toString(parser.getPoints()));
 
               plugin.getServer().dispatchCommand(plugin.getServer().getConsoleSender(), actionCommand);
             }
